feat(file-input): add file input component
This commit deprecates the previous `.clr-file-wrapper` styles due to
design changes. The component uses separate `.clr-file-input-wrapper`
styles.

There is a visual bug with the helper and error/success text in the
compact layout. This is a existing bug that exists will all of the form
controls. We will address this separately.

CDE-1668